+++ packages/cojson/CHANGELOG.md
@@ -1,5 +1,17 @@
 # cojson
 
+## 0.17.2
+
+### Patch Changes
+
+- 5b2b16a: Filter invalid sessions when storing a content message
+
+## 0.17.1
+
+### Patch Changes
+
+- 2fd88b9: Add debug info to sync correction errors
+
 ## 0.17.0
 
 ## 0.16.6